refactor: PdfUploadModule > style 디렉토리 삭제
내부에 있는 파일은 style.ts로 변경하여 PdfUploadModule
디렉토리로(기존에서 상위로) 이동하고 style 디렉토리는 삭제
