+++ src/components/PdfUploadModule/style.ts
